feat: level별로 랜덤 데이터 조회
