+++ src/components/chatBot/aiTutorHome/TodayLearn.tsx
@@ -1,12 +1,11 @@
 "use client";
 
 import { createClient } from "@/utils/supabase/client";
-import { useCallback } from "react";
-
 import { useRouter } from "next/navigation";
 import { format } from "date-fns";
 import { useMutation, useQuery } from "@tanstack/react-query";
 import { Tables } from "../../../../database.types";
+import { reviewApi } from "@/services/supabaseChatbot";
 
 type SituationType = Tables<"situation">;
 
@@ -15,46 +14,15 @@ const TodayLearn = () => {
   const router = useRouter();
 
   // 유저 정보 조회
-  const getUserInfo = async () => {
-    const {
-      data: { user },
-      error
-    } = await supabase.auth.getUser();
-    if (error) {
-      console.log("유저 정보를 가져오는 데에 실패하였습니다! ", error);
-      return null;
-    }
-    return user;
-  };
-
   const { data: user } = useQuery({
     queryKey: ["userInfo"],
-    queryFn: getUserInfo
+    queryFn: reviewApi.getUserInfo
   });
 
   // situation 조회
-  const getSituations = useCallback(async () => {
-    try {
-      const { data, error } = await supabase.from("situation").select("*");
-
-      if (error) {
-        throw error;
-      }
-
-      if (data) {
-        // 데이터가 3개 이상일 경우 랜덤으로 3개 선택
-        const randomSiuations = data.sort(() => 0.5 - Math.random()).slice(0, 3);
-        return randomSiuations;
-      }
-    } catch (error) {
-      console.log("situation을 가져오는 데에 실패하였습니다!", error);
-      throw error;
-    }
-  }, [supabase]);
-
   const { data: situations } = useQuery({
     queryKey: ["situations"],
-    queryFn: getSituations
+    queryFn: reviewApi.getEachLevel
   });
 
   // review 테이블에 유저가 선택한 학습 추가
